% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFormat_Utils.R
\name{combine.data.frame}
\alias{combine.data.frame}
\title{merge dataframes}
\usage{
combine.data.frame(df1, df2, fill=NA, merge.cols=TRUE)
}
\arguments{
\item{df1}{a dataframe}

\item{df2}{a dataframe}

\item{fill}{character or NA. A value to put into the cells that have no data. Default NA}

\item{merge.cols}{boolean. Merge colomns with common name. Default TRUE}
}
\value{
a data.frame
}
\description{
combine.data.frame merges two dataframes, completing the rows and columns that are not shared by the dataframes.
}
\details{
Columns with matching names can or cannot be merged, rows are automatically bound (a wrapper of rbind), not merged. Any missing data as a result of the non-matchng columns will be filled by the fill argument.
}
\author{
Maxime Sweetlove ccBY 4.0 2019
}
