% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQC_Utils.R
\name{dataQC.taxaNames}
\alias{dataQC.taxaNames}
\title{perform a quality control onf taxonomic names}
\usage{
dataQC.taxaNames(taxaNames)
}
\arguments{
\item{taxaNames}{character vector. a list of scientific taxonomic names}
}
\value{
a vector with the checked taxon names
}
\description{
checks a list of taxonomic names, and perform a basic quality control
}
\details{
looks for trailing spaces or common typos in taxonomic names.
}
\author{
Maxime Sweetlove CC-BY 4.0 2020
}
