% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQC_Utils.R
\name{dataQC.TaxonListFromData}
\alias{dataQC.TaxonListFromData}
\title{extract the finest resolution taxonomy from a dataset}
\usage{
find.sampleTaxon(dataset)
}
\arguments{
\item{dataset}{a data.frame. The dataset with samples as rows, and taxonomy information in the columns. using the MIxS or DarwinCore taxonomy terms, the taxonomy information will be extracted}
}
\value{
a vector with the highest level taxonomic name found, with genus and species epithet separated by a space.
}
\description{
tries to find taxonomic names for samples (rows) in a dataset (data.frame),
}
\author{
Maxime Sweetlove CC-BY 4.0 2020
}
