% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Utils.R
\name{get.ENAName}
\alias{get.ENAName}
\title{convert a MIxS term into it's ENA variant}
\usage{
get.ENAName(variable)
}
\arguments{
\item{variable}{character a MIxS term.}
}
\value{
a character
}
\description{
get the ENA variant a MIxS term
}
\author{
Maxime Sweetlove ccBY 4.0 2019
}
