% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobeDataTools.R
\docType{package}
\name{MicrobeDataTools}
\alias{MicrobeDataTools}
\title{MicrobeDataTools: A package with tools to format and standardize microbial 'omics datasets.}
\description{
This package provides 5 cathegoties of tools, these are:
formating functions, standardization functions, quality control functions, data archiving functions and downloading data functions
In addition, there are also different libraries with terms of the MIxS and DarwinCore standards, term variants, synonyms and translations.
}
\section{Classes}{

MIxS.metadata
DwC.event
DwC.occurrence
}

\section{Libraries}{

TaxIDLib
TermsLib
TermsSyn
TermsSyn_DwC
MarsLib
ENA_allowed_terms
}

\section{Functions}{

check.valid.metadata.DwC
check.valid.metadata.MIxS  
combine.data
combine.data.frame
commonTax.to.NCBI.TaxID
coordinate.to.decimal
dataQC.completeTaxaNamesFromRegistery
dataQC.dateCheck
dataQC.DwC
dataQC.DwC_general
dataQC.eventStructure
dataQC.findNames
dataQC.generate.footprintWKT
dataQC.guess.env_package.from.data
dataQC.LatitudeLongitudeCheck
dataQC.MIxS
dataQC.taxaNames
dataQC.TaxonListFromData
dataQC.TermsCheck
download.sequences.INSDC
eMoF.to.wideTable
ENA_checklistAccession
ENA_geoloc
ENA_instrument
ENA_select
ENA_strat
FileNames.to.Table
find.dataset
get.BioProject.metadata.INSDC
get.boundingBox
get.ENAName
get.insertSize
get.sample.attributes.INSDC
multi.warnings
prep.metadata.ENA
rename.sequenceFiles
sync.metadata.sequenceFiles
term.definition
wideTab.to.hierarchicalTab
wideTable.to.eMoF
}

