% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFormat_Utils.R
\name{wideTable.to.eMoF}
\alias{wideTable.to.eMoF}
\title{convert dataframe to a DarwinCore extended Measurement or Fact (eMoF) file}
\usage{
wideTable.to.eMoF(dataset)
}
\arguments{
\item{metadata.object}{a metadata.MIxS object}

\item{variables}{a character vector. a list of the variables that need to be included in the eMoF}
}
\value{
a data.frame formatted as an extended Measurement or Fact table
}
\description{
converts a dataframe to a DarwinCore extended Measurement Or Fact (eMOF) file
}
\author{
Maxime Sweetlove ccBY 4.0 2020
}
